requirements.txt
aiosqlite==0.18.0
anyio==3.6.2
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
click==8.1.3
coverage==7.1.0
ecdsa==0.18.0
exceptiongroup==1.1.0
fastapi==0.89.1
gprof2dot==2022.7.29
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
mypy==1.0.0
mypy-extensions==1.0.0
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
pyasn1==0.4.8
pydantic==1.10.4
pytest==7.2.1
pytest-asyncio==0.20.3
pytest-cov==4.0.0
pytest-profiling==1.7.0
python-dotenv==0.21.1
python-jose==3.3.0
python-multipart==0.0.5
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.2
sqlalchemy-stubs==0.4
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
uvicorn==0.20.0
